revise qsdr code

+++ libs/dsi/basic_voxel.hpp
@@ -103,7 +103,7 @@ public:// used in QSDR
     bool output_rdi = false;
     bool qsdr = false;
     tipl::vector<3,int> csf_pos1,csf_pos2,csf_pos3,csf_pos4;
-    double R2;
+    float R2;
 public: // for QSDR associated T1WT2W
     std::vector<tipl::image<float,3> > other_image;
     std::vector<std::string> other_image_name;
